Fix remaining stuff of d5edd95. * Forgot to remove some `#endifs`. [datadriven] * Fix style warnings. [quadrature]

+++ quadrature/tests/test_quadrature.cpp
@@ -123,7 +123,7 @@ BOOST_AUTO_TEST_CASE(testOperationMCAdvanced) {
     sgpp::op_factory::createOperationQuadrature(*grid));
   double analyticIntegral = opQuad->doQuadrature(alpha);
 
-  BOOST_CHECK_CLOSE(analyticIntegral, analyticResult, double(1e-12) );
+  BOOST_CHECK_CLOSE(analyticIntegral, analyticResult, 1e-12);
 
   std::vector<size_t> blockSize(dim);
